pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dazllm"
version = "0.2.0"
description = "Simple, unified interface for all major LLMs"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Darren Oakey", email = "darren.oakey@insidemind.com.au"}
]
keywords = ["llm", "ai", "openai", "anthropic", "claude", "gemini", "ollama", "chatgpt", "gpt-4"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "keyring>=24.0.0",
    "requests>=2.25.0",
    "colorama>=0.4.0",
    "pydantic>=2.0.0",
    "jsonschema>=4.0.0",
    "Pillow>=9.0.0",
    "openai>=1.0.0", 
    "anthropic>=0.3.0",
    "google-generativeai>=0.3.0",
    "google-ai-generativelanguage",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/darrenoakey/dazllm"
Documentation = "https://github.com/darrenoakey/dazllm#readme"
Repository = "https://github.com/darrenoakey/dazllm.git"
"Bug Reports" = "https://github.com/darrenoakey/dazllm/issues"

[project.scripts]
dazllm = "dazllm.cli:main"
